login split && readme
