+++ .gitignore
@@ -1,3 +1,4 @@
 go.sum
 .env
 mysql-data
+vendor
