+++ README.md
@@ -1,0 +1,44 @@
+# Website-Structure
+
+## Badges
+
+![GitHub go.mod Go version](https://img.shields.io/github/go-mod/go-version/Keanus-In-Reevesverse/website-structure)
+
+## Index 
+
+* [Index](#index)
+* [Badges](#badges)
+* [Project Description](#project-description)
+* [Project Status](#project-status)
+* [Tecnologies](#tecnologies)
+* [Running](#running)
+* [Developers](#developers)
+
+## Project Description
+
+<p align="center">A website backend structure made with golang-gin</p>
+
+## Project Status
+
+:construction: Initial Stages :construction:
+
+## Tecnologies
+
+- ``Golang 1.18.2``
+- ``Gin Framework``
+- ``Docker``
+- ``MySQL``
+- ``Swagger``
+
+## Running
+
+Follow .envexample for database connection and port definitions
+
+To run this project just do a:
+`go run main.go`
+
+To run tests just do a:
+`go test`
+
+To see living documentation access:
+`localhost:8080/swagger/index.html`
