+++ models/login.go
@@ -1,0 +1,7 @@
+package models
+
+type Login struct {
+	UserId   int64
+	Email    string
+	Password string
+}
